requirements.txt
openai>=1.40.0
google-generativeai>=0.3.0
anthropic>=0.18.0
pydantic>=2.6.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
PyYAML>=6.0.0
requests>=2.31.0
rich>=13.7.0
fastapi>=0.111.0
uvicorn>=0.30.0
pandas>=2.1.0
streamlit>=1.28.0
plotly>=5.17.0
pillow>=10.1.0
playwright>=1.40.0
